backend/requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
Pillow>=10.0.0
httpx>=0.26.0
